refactor(media): extract nika-media crate + fix clippy workspace-wide
Extract CAS store, media detection, processor, types, and error
to standalone nika-media crate. 3469 lines, 120 tests.

- CasStore with blake3 hashing and compression (media-compression feature)
- MediaProcessor with inline safe image decoding
- Forward media-thumbnail and media-compression features to nika-media
- Fix 10+ clippy warnings across nika-lsp and nika-lsp-core

+++ tools/nika-lsp-core/src/handlers/completion.rs
@@ -916,8 +916,8 @@ mod tests {
     /// Helper: detect context and compute completions.
     fn complete_at(text: &str, line: usize, character: usize) -> Vec<CompletionItem> {
         let offset = text_offset(text, line, character);
-        let ctx = detect_context(text, offset as u32, None);
-        completions(text, offset as u32, &ctx)
+        let ctx = detect_context(text, offset, None);
+        completions(text, offset, &ctx)
     }
 
     /// Convert line/character to byte offset.
